Initial introduction of ThreadPools
In order to dedicate some threads to prefetching, we need a way of only
assigning tasks to a subset of threads.

To do this, we introduce ThreadPools, which is kind of a generalization
of threading infrastructure.

Ideally, we'll be able to have one thread pool handle prefetching, and
another deal with performing queries.