Swap Euclidean implementation.
Implementing Euclidean points as wrappers around NTuples was beginning
to get cumbersome and annoying.

Here, the implementation is swapped out for using SVectors (which
themselves are technically wrappers around tuples, but carry much more
convenient behavior than tuples).

Hopefully, we'll be able to retrofit some of the PQ code to be cleaner,
and more importantly faster.