Initial implementation of much faster PQ encoding.
At the moment, the time it takes to convert a graph into the PQ expanded
format is unreasonably long.

This is in initial implementation using explicit avx instructions to
greatly accelerate the time it takes to encode points.

For reference, the time before this optimization is around 25
microseconds per encoding. After these changes, it's around 1.5
microseconds.