Changes to prefetching and quantization.
Shelving quantization at the moment, I need to stew about how to cleanly
express computations so we get enough efficiency.

Switched back to prefetching and have a first working prototype.
Speedup on Sift100M with graph and data in PM is around 8-9%, so not
amazing, but also not actually terrible.