Update PQ.
This commit does a couple of things. First, it brings the PQ package in
`contrib/` up to date with the main GraphANN code.

Further, it modifies many of the hooks in the search algorithms to
facilitate packages like PQ. This includes custom finalizer passes to
allow reranking, and an update to the exhaustive search that allows us
to change the iteration order over the dataset.